remove circ dep
